requirements.txt
Flask==3.1.0
huggingface_hub==0.26.0
numpy<2.0.0 
pandas==2.2.0
pydub==0.25.1
python-dotenv==1.0.1
soundfile==0.12.1
Werkzeug==3.1.3
pyarrow==14.0.1
fastparquet
waitress
jinja2
apscheduler
numpy
pocketbase
flask-caching
filelock
apscheduler
pytz
sqlalchemy
psycopg2-binary
psutil
PyJWT
flask-compress
flask_login
scipy==1.14.1